+++ packages/integration/lib/src/bloc/comment/comment_handler/handler/ongoing_comment_event.dart
@@ -5,6 +5,7 @@ part of '../comment_handler_bloc.dart';
 /// i.e. A user is editing a comment/reply
 /// [LMFeedCommentOngoingEvent] is used to send the request to the handler
 /// or replying to a existing comment
+/// {@endtemplate}
 void handleOngoingCommentEvent(LMFeedCommentOngoingEvent event,
         Emitter<LMFeedCommentHandlerState> emit) =>
     emit(