+++ packages/integration/lib/src/bloc/post/post_event.dart
@@ -58,16 +58,3 @@ class LMFeedUpdatePostEvent extends LMFeedPostEvents {
   @override
   List<Object> get props => [post, DateTime.now().millisecondsSinceEpoch];
 }
-
-class LMFeedTogglePinPostEvent extends LMFeedPostEvents {
-  final String postId;
-  final bool isPinned;
-
-  LMFeedTogglePinPostEvent({
-    required this.postId,
-    required this.isPinned,
-  });
-
-  @override
-  List<Object> get props => [postId, isPinned];
-}