+++ packages/ui/lib/src/widgets/common/icon/icon.dart
@@ -12,7 +12,8 @@ class LMFeedIcon extends StatelessWidget {
   /// if [LMFeedIconType.icon] then pass icon of type [IconData]
   final IconData? icon;
 
-  /// if [LMFeedIconType.png] or [LMFeedIconType.svg] then pass path of icon [String]
+  /// if [LMFeedIconType.png] or [LMFeedIconType.svg] then 
+  /// pass path of icon [String]
   final String? assetPath;
 
   /// style class for styling the icon [LMFeedIconStyle]
@@ -41,12 +42,9 @@ class LMFeedIcon extends StatelessWidget {
           height: style.size?.abs() ?? 24,
           child: SvgPicture.asset(
             assetPath!,
-            colorFilter: style.color == null
-                ? null
-                : ColorFilter.mode(
-                    style.color!,
-                    BlendMode.srcATop,
-                  ),
+            color: style.color,
+            colorBlendMode:
+                style.color == null ? BlendMode.srcIn : BlendMode.srcATop,
             fit: style.fit ?? BoxFit.contain,
           ),
         );