+++ packages/ui/lib/src/widgets/post/post_header.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:likeminds_feed_flutter_ui/src/models/models.dart';
 import 'package:likeminds_feed_flutter_ui/src/utils/index.dart';
-import 'package:likeminds_feed_flutter_ui/src/utils/feed_time_ago/feed_time_ago.dart';
 import 'package:likeminds_feed_flutter_ui/src/widgets/widgets.dart';
 
 class LMFeedPostHeader extends StatelessWidget {