+++ packages/ui/CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+## 1.0.0
 
-* TODO: Describe initial release.
+Initial release of LikeMinds Feed Flutter UI library
+
+## 1.0.1
+
+Patch update wuth dependency update
+
+## 1.0.2
+
+Migrating to Flutter v3.10.6